requirements.txt
transformers
torch
Pillow 